serverlist

+++ src/components/ServersList.jsx
@@ -19,7 +19,7 @@ export default function ServersList() {
     console.log(ServerList, ServerList.length)
     return (
         <div className="ServersList">
-            {ServerList.length != 0 ? ServerList.map((s) => { return <ServerCard key={s.id} serverName={s.name} serverPrice={parseInt(s.price)} serverFeature={[s.features.cores, parseInt(s.features.core_frequency).toFixed(1), s.features.ram, s.features.ram_type, s.features.disk, s.features.disk_type, parseInt(s.features.network_speed)]} /> }) : <p>Сервера не найдены!</p>}
+            {ServerList.length != 0 ? ServerList.map((s) => { return <ServerCard key={s.id} serverName={s.name} serverPrice={parseInt(s.price)} serverFeature={[s.features.cores, parseFloat(s.features.core_frequency).toFixed(1), s.features.ram, s.features.ram_type, s.features.disk, s.features.disk_type, parseInt(s.features.network_speed)]} /> }) : <p>Сервера не найдены!</p>}
         </div>
     )
 }
